+++ modal-js/src/image.ts
@@ -9,6 +9,9 @@ import { client } from "./client";
 import { App } from "./app";
 import { Secret } from "./secret";
 import { imageBuilderVersion } from "./config";
+import { ClientError } from "nice-grpc";
+import { Status } from "nice-grpc";
+import { NotFoundError } from "./errors";
 
 /** Options for deleting an Image. */
 export type ImageDeleteOptions = Record<never, never>;
@@ -38,8 +41,21 @@ export class Image {
    *
    * @param imageId - Image ID.
    */
-  static fromId(imageId: string): Image {
-    return new Image(imageId, "");
+  static async fromId(imageId: string): Promise<Image> {
+    try {
+      const resp = await client.imageFromId({ imageId });
+      return new Image(resp.imageId, "");
+    } catch (err) {
+      if (err instanceof ClientError && err.code === Status.NOT_FOUND)
+        throw new NotFoundError(err.details);
+      if (
+        err instanceof ClientError &&
+        err.code === Status.FAILED_PRECONDITION &&
+        err.details.includes("Could not find image with ID")
+      )
+        throw new NotFoundError(err.details);
+      throw err;
+    }
   }
 
   /**
@@ -187,8 +203,11 @@ export class Image {
   }
 
   /** Delete an Image by ID. Warning: This removes an *entire Image*, and cannot be undone. */
-  static async delete(name: string, _: ImageDeleteOptions = {}): Promise<void> {
-    const image = await Image.fromId(name);
+  static async delete(
+    imageId: string,
+    _: ImageDeleteOptions = {},
+  ): Promise<void> {
+    const image = await Image.fromId(imageId);
     await client.imageDelete({ imageId: image.imageId });
   }
 }