+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
 ## Unreleased
 
 - Added `Image.delete()` (JS) and `ImageDelete()` (Go).
+- Changed `Image.fromId()` (JS) and `NewImageFromId()` (Go) to throw a `NotFoundError` if the Image does not exist.
 
 ## modal-js/v0.3.19, modal-go/v0.0.19